pyproject.toml
[project]
name = "macos-host-dashboard"
version = "0.1.0"
description = "Local MacOS host monitoring with Prometheus and Grafana"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
